Add id to each task and append task to the tail of the json file

+++ CLI_Task_Manager_JSON.py
@@ -1,5 +1,6 @@
+# import datetime
 import json
-import datetime #TODO I'll add timestamp to user's input
+import os
 import re
 
 user_input= input("Please enter your task, task's priority, task's deadline seperate them with ',': ")
@@ -12,6 +13,23 @@ def take_task(user_task):
 
 print(f"User's task information is {take_task(user_input)}.")
 
-with open('/tmp/tasks.json', 'w') as file:
-    data = take_task(user_input)
-    json.dump(data, file)
+if not os.path.exists('/tmp/tasks.json'):
+    with open('/tmp/tasks.json', 'w') as file:
+        data = take_task(user_input)
+        data["id"] = 1
+        json.dump([data], file, indent=4, sort_keys=True)
+
+else:
+    with open('/tmp/tasks.json', 'r') as file:
+        data = json.load(file)
+        next_id = max((item['id'] for item in data), default=0) + 1
+
+        new_entry = take_task(user_input)
+        new_entry["id"] = next_id
+
+        data.append(new_entry)
+        
+    with open('/tmp/tasks.json', 'w') as file:
+        json.dump(data, file, indent=4, sort_keys=True)
+
+
